(Refact) Restructure file system for better organization
