+++ src/App.tsx
@@ -1,8 +1,10 @@
 import { RouterProvider } from 'react-router-dom';
 import { router } from '@/router';
 import { useEffect } from 'react';
-import { useAuthStore } from '@/store/auth';
+import { useAuthStore } from '@/features/Member/store/auth';
 // import { useGlobalErrorToast } from '@/hooks/useGlobalErrorToast';
+// import { useViewportHeight } from './hooks/useViewportHeight';
+// import { AnimatePresence } from 'motion/react';
 
 function App() {
   // useGlobalErrorToast();
