+++ src/features/Company/components/CompanyTitle.tsx
