+++ src/features/Map/assets/clusterStyles.ts
