+++ src/features/Map/assets/company-card-background.png
