+++ src/features/Map/assets/map.ts
