+++ src/features/Member/lib/axios.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosError } from 'axios';
-import { useAuthStore } from '@/store/auth';
+import { useAuthStore } from '@/features/Member/store/auth';
 import { publishErrorToast } from '@/lib/errorEvents'
-import { logout } from '@/lib/logout';
+import { logout } from '@/features/Member/lib/logout';
 
 export const instance = axios.create({
   baseURL: import.meta.env.VITE_API_BASE_URL, // .env
