+++ src/hooks/useToggleBookmarkMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
-import { instance as axios } from '@/lib/axios';
-import { useAuthStore } from '@/store/auth';
+import { instance as axios } from '@/features/Member/lib/axios';
+import { useAuthStore } from '@/features/Member/store/auth';
 
 export const useToggleBookmarkMutation = ({
 //   token,
