+++ src/pages/LoginRequired.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@/components/ui/button';
-import { instance as axios } from '@/lib/axios';
+import { instance as axios } from '@/features/Member/lib/axios';
 
 export default function LoginRequired() {
   const handleLogin = async () => {
